Added fixes to ccoreconf remove warnings

+++ examples/demo_functionalities.c
@@ -89,7 +89,7 @@ void main() {
     /*Find the nodes corresponding to a String Characteristics and specific keys*/
 
     // keys MUST be initialized properly and must be NON Empty
-    int64_t keys[] = {5, 3, 1000068, 1, 1000018, 0};
+    int keys[] = {5, 3, 1000068, 1, 1000018, 0};
     size_t keyLength = sizeof(keys) / sizeof(keys[0]);
     
     // Build a valid SidIdentifierT object and then call traverseCORECONFWithKeys
